Consolidate debug-mock into cmd/debug via --mock flag
Removes the separate cmd/debug-mock binary and adds a -mock flag to
cmd/debug instead. When passed, it skips config.Load() and wires up
inline fake implementations for all external dependencies.
